# application-config-service/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Application Config Service API
  description: |
    RESTful API for managing application configurations and modules in the Onified platform.
    
    This service provides endpoints for:
    - Creating, reading, updating, and deleting applications
    - Managing application modules
    - Application configuration management
    
    ## Features
    - **Application Management**: Full CRUD operations for applications
    - **Module Management**: Create and manage modules within applications
    - **Active/Inactive Status**: Toggle application and module status
    - **Standardized Responses**: Consistent API response format
    
    ## Authentication
    This service requires proper authentication and authorization.
    
    ## Rate Limiting
    API calls are subject to rate limiting to ensure service stability.
  version: 1.0.0
  contact:
    name: Onified Development Team
    email: dev@onified.ai
    url: https://onified.ai
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:9082
    description: Local Development Server
  - url: https://api.onified.ai
    description: Production Server
...
paths:
...
  /api/modules:
    post:
      tags:
        - Module Management
      summary: Create a new application module
      description: Creates a new module for an existing application. The module code must be unique within the application.
      operationId: createAppModule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModuleRequestDTO'
            examples:
              create-module:
                summary: Create Module Example
                value:
                  appCode: "APP001"
                  moduleCode: "MOD001"
                  isActive: true
      responses:
        '201':
          description: Module created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ModuleResponseDTO'
              examples:
                success:
                  summary: Success Response
                  value:
                    statusCode: 201
                    status: "SUCCESS"
                    body:
                      moduleId: 1
                      appCode: "APP001"
                      moduleCode: "MOD001"
                      isActive: true
        '400':
          description: Bad request - Invalid input data
        '404':
          description: Application not found
        '409':
          description: Conflict - Module code already exists in the application
        '500':
          description: Internal server error
  /api/modules/{moduleId}:
    get:
      tags:
        - Module Management
      summary: Get module by ID
      description: Retrieves a module by its unique module ID.
      operationId: getAppModuleById
      parameters:
        - name: moduleId
          in: path
          required: true
          description: Unique module identifier
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Module found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ModuleResponseDTO'
        '404':
          description: Module not found
        '500':
          description: Internal server error
    put:
      tags:
        - Module Management
      summary: Update an application module
      description: Updates an existing module with new details. The module ID in the path must match the one being updated.
      operationId: updateAppModule
      parameters:
        - name: moduleId
          in: path
          required: true
          description: Module ID to update
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModuleRequestDTO'
            examples:
              update-module:
                summary: Update Module Example
                value:
                  appCode: "APP001"
                  moduleCode: "MOD001"
                  isActive: false
      responses:
        '200':
          description: Module updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ModuleResponseDTO'
        '400':
          description: Bad request - Invalid input data
        '404':
          description: Module not found
        '500':
          description: Internal server error
    delete:
      tags:
        - Module Management
      summary: Delete an application module
      description: Deletes a module by its module ID. This operation is irreversible.
      operationId: deleteAppModule
      parameters:
        - name: moduleId
          in: path
          required: true
          description: Module ID to delete
          schema:
            type: integer
            example: 1
      responses:
        '204':
          description: Module deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_String'
              examples:
                success:
                  summary: Success Response
                  value:
                    statusCode: 204
                    status: "SUCCESS"
                    body: "Module deleted successfully."
        '404':
          description: Module not found
        '500':
          description: Internal server error
  /api/modules/by-app/{appCode}:
    get:
      tags:
        - Module Management
      summary: Get modules by application code
      description: Retrieves all modules for a specific application.
      operationId: getAppModulesByAppCode
      parameters:
        - name: appCode
          in: path
          required: true
          description: Application code to get modules for
          schema:
            type: string
            example: "APP001"
      responses:
        '200':
          description: Modules retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_List_ModuleResponseDTO'
        '404':
          description: Application not found
        '500':
          description: Internal server error
...
components:
  schemas:
...
    ModuleRequestDTO:
      type: object
      description: Request DTO for creating or updating an application module
      required:
        - appCode
        - moduleCode
        - isActive
      properties:
        appCode:
          type: string
          description: Application code that this module belongs to
          example: "APP001"
          minLength: 1
          maxLength: 50
        moduleCode:
          type: string
          description: Unique module code identifier
          example: "MOD001"
          minLength: 1
          maxLength: 50
        isActive:
          type: boolean
          description: Whether the module is active
          example: true
    ModuleResponseDTO:
      type: object
      description: Response DTO for application module data
      properties:
        moduleId:
          type: integer
          description: Unique module identifier
          example: 1
        appCode:
          type: string
          description: Application code that this module belongs to
          example: "APP001"
        moduleCode:
          type: string
          description: Unique module code identifier
          example: "MOD001"
        isActive:
          type: boolean
          description: Whether the module is active
          example: true
...
    ApiResponse_ModuleResponseDTO:
      type: object
      description: Standardized API response wrapper for ModuleResponseDTO
      properties:
        statusCode:
          type: integer
          description: HTTP status code
          example: 200
        status:
          type: string
          description: Status message
          example: "SUCCESS"
        body:
          $ref: '#/components/schemas/ModuleResponseDTO'
...
    ApiResponse_List_ModuleResponseDTO:
      type: object
      description: Standardized API response wrapper for list of ModuleResponseDTO
      properties:
        statusCode:
          type: integer
          description: HTTP status code
          example: 200
        status:
          type: string
          description: Status message
          example: "SUCCESS"
        body:
          type: array
          items:
            $ref: '#/components/schemas/ModuleResponseDTO'
    ApiResponse_String:
      type: object
      description: Standardized API response wrapper for String
      properties:
        statusCode:
          type: integer
          description: HTTP status code
          example: 204
        status:
          type: string
          description: Status message
          example: "SUCCESS"
        body:
          type: string
          description: Response body containing the actual data
          example: "Application deleted successfully."
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
security:
  - bearerAuth: [] 